Added heuristics.potential, the bets yet
heuristics.potential is the sum of all winning scenarios that are now
unachievable to the opponent, minus all winning scenarios that are now
unachieavable to you.

This heuristic is the best yet, as it is able to reward "occupying the
middle", which the previous heuristics could not do.

On the other hand, it overemphasizes the point, and some tweaking
should be done to make it strive more for the goal itself.

Still have to implement alpha-beta pruning.