Created prototype for machine learning model
Simple model for machine learning-based play. lib/brain defines a class that is
able to play and learn over time given results of play. lib/train is a script for
training said class.

Model is very rudimentary, arbitrary, and time-consuming to train. It is also
space-inefficient and takes a dumb approach based on "which move have I played in
this scenario in the past and got good results with?"